requirements.txt
langchain
langchain_core
langchain-community
pymupdf
sentence-transformers  #embedding model
chromadb
langchain-chroma
langchain-huggingface 
huggingface_hub
python-dotenv
streamlit
nbimporter